rename JotPad.py to main.py
